Move library functions to new operator package
except the code that is originally generated by the operator-sdk cli
