+++ controllers/schedule_it_test.go
@@ -10,10 +10,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	k8upv1alpha1 "github.com/vshn/k8up/api/v1alpha1"
-	"github.com/vshn/k8up/cfg"
 	"github.com/vshn/k8up/controllers"
-	"github.com/vshn/k8up/handler"
-	"github.com/vshn/k8up/scheduler"
+	"github.com/vshn/k8up/operator/cfg"
+	"github.com/vshn/k8up/operator/handler"
+	"github.com/vshn/k8up/operator/scheduler"
 )
 
 type (
