+++ operator/cfg/config_test.go
