+++ operator/executor/backup_utils.go
@@ -8,8 +8,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/vshn/k8up/cfg"
-	"github.com/vshn/k8up/observer"
+	"github.com/vshn/k8up/operator/cfg"
+	"github.com/vshn/k8up/operator/observer"
 )
 
 func (b *BackupExecutor) fetchPVCs(list client.ObjectList) error {
