+++ operator/executor/cleaner/cleaner.go
@@ -5,14 +5,14 @@ import (
 	"fmt"
 	"sort"
 
-	k8upv1alpha1 "github.com/vshn/k8up/api/v1alpha1"
-	"github.com/vshn/k8up/cfg"
-	"github.com/vshn/k8up/job"
-
 	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	k8upv1alpha1 "github.com/vshn/k8up/api/v1alpha1"
+	"github.com/vshn/k8up/operator/cfg"
+	"github.com/vshn/k8up/operator/job"
 )
 
 // ObjectCleaner cleans old, finished job objects.
