+++ operator/handler/job.go
@@ -9,8 +9,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/vshn/k8up/api/v1alpha1"
-	"github.com/vshn/k8up/job"
-	"github.com/vshn/k8up/observer"
+	"github.com/vshn/k8up/operator/job"
+	"github.com/vshn/k8up/operator/observer"
 )
 
 const (
