+++ operator/handler/randomizer_test.go
