+++ operator/observer/observer_test.go
